+++ packages/orq-rc/RELEASES.md
@@ -10949,3 +10949,13 @@ Based on:
 - [python v4.15.0-rc.48] packages/orq-rc
 ### Releases
 - [PyPI v4.15.0-rc.48] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.48 - packages/orq-rc
+
+## 2026-09-15 10:26:42
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.797.0 (2.937.18) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.15.0-rc.49] packages/orq-rc
+### Releases
+- [PyPI v4.15.0-rc.49] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.49 - packages/orq-rc